App/Form_Comparator_graybox6nn/requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
pdf2image==1.16.3
opencv-python==4.9.0.80
numpy==1.26.4
pytesseract==0.3.10
Pillow==10.2.0 